requirements.txt
yfinance
rich
prettytable
requests
pandas
talib
xmltodict
pytz
matplotlib
Pillow
newsapi-python
streamlit